Create one file with all GUI handlers
