+++ pytripgui/field_vc/field_view.py
@@ -1,17 +1,8 @@
-from PyQt5 import QtWidgets, uic
+from pytripgui.view.qt_gui import UiFieldDialog
 from PyQt5.QtCore import Qt
-from pytripgui import util
-import os
 
 import logging
 logger = logging.getLogger(__name__)
-
-
-class UiFieldDialog(QtWidgets.QDialog):
-    def __init__(self):
-        super(UiFieldDialog, self).__init__()
-        ui_path = os.path.join(util.main_dir(), 'view', 'field.ui')
-        uic.loadUi(ui_path, self)
 
 
 class FieldQtView(object):
