+++ pytripgui/plan_vc/plan_view.py
@@ -1,18 +1,9 @@
-from PyQt5 import QtWidgets, uic
+from pytripgui.view.qt_gui import UiPlanDialog
 from PyQt5.QtCore import Qt
 from PyQt5.QtWidgets import QTreeWidgetItem
-from pytripgui import util
-import os
 
 import logging
 logger = logging.getLogger(__name__)
-
-
-class UiPlanDialog(QtWidgets.QDialog):
-    def __init__(self):
-        super(UiPlanDialog, self).__init__()
-        ui_path = os.path.join(util.main_dir(), 'view', 'plan.ui')
-        uic.loadUi(ui_path, self)
 
 
 class PlanQtView(object):
