--- pytripgui/view/gen/__init__.py
@@ -1,16 +1,0 @@
-"""
-    This file is part of pytripgui.
-
-    pytripgui is free software: you can redistribute it and/or modify
-    it under the terms of the GNU General Public License as published by
-    the Free Software Foundation, either version 3 of the License, or
-    (at your option) any later version.
-
-    pytripgui is distributed in the hope that it will be useful,
-    but WITHOUT ANY WARRANTY; without even the implied warranty of
-    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-    GNU General Public License for more details.
-
-    You should have received a copy of the GNU General Public License
-    along with pytripgui.  If not, see <http://www.gnu.org/licenses/>
-"""
